Filter non-biallelic / haplotype-style variants in a radiator GDS

Filter non-biallelic / haplotype-style variants in a radiator GDS

Usage

filter_haplotype_format(
  gds,
  filters.parameters,
  path.folder,
  file.date,
  filter.haplotype.format = FALSE,
  internal = TRUE,
  verbose = TRUE
)

Arguments

gds

A radiator GDS object (SeqVarGDSClass) with a /radiator/markers.meta node containing at least MARKERS, REF and ALT.

filters.parameters

The current filter-parameter object generated by genome_parameters().

path.folder

(character) Root folder used to store filter outputs.

file.date

(character) Timestamp string (e.g. "20251203@1919") used in filenames.

filter.haplotype.format

(logical) Activate or deactivate this filter. Default: filter.haplotype.format = FALSE.

internal

(logical) Indicate if the function is called internally by other radiator functions, used for controlling side-effect messages and paths. Default: internal = TRUE.

verbose

(logical) When TRUE, the function prints messages about detected haplotypes and written files. Default: verbose = TRUE.

Value

A list with:

markers.meta – updated markers metadata tibble;

filters.parameters – updated filter-parameter object.